Add files via upload
In order to make an efficient machine learning model with fairly good predicting power , EDA is the basic pre step.

EDA is a very very important step without which we can't proceed to model building

1. EDA -Day1 - it has the basic and self explanatory  code using which we can import our data from csv or excel format into python  , make it readable using Data Frames and perform various functions

2. EDA - Day2 - it consist of basic knowledge which we can extract from data eg. mean , median , mode whether the data is skewed or not , indexing, replacing, renaming , reshaping etc.

3. EDA - Day3 - How to clean data ,  deal with any missing values , how to impute missing values , plotting data , graphs charts etc

4. EDA -Day4 - How to normalise data , deal with outliers , concept of zscore, feature engineering.

For performing EDA we use pandas package from python library

Pandas is a high-level data manipulation tool developed by Wes McKinney. It is built on the Numpy package and its key data structure is called the DataFrame. DataFrames allow you to store and manipulate tabular data in rows of observations and columns of variables.